Fix jumping after touch zooming also for mobile clients of normal Online
Don't check window.ThisIsAMobileApp but instead use a much more
convoluted test to catch only the cases where we do want the document
to be scrolled, and ignore the case that causes the irritating
jumping.

In _onUpdateCursor() in TileLayer.js, we know that the scroll
parameter is not undefined when and only when _onUpdateCursor() is
called from _onInvalidateCursorMsg(). We pass that information down to
the onUpdate callback handler passed to the _tweenTo() function in the
_scrollTo() function in jquery.mCustomScrollbar.js where it shows up
as property of the options object.

In that location we also make use of the incidental fact that
options.timeout==1 only when that code is called from mouse-wheel scrolling.

Quite possibly all this could be done much cleaner.

Change-Id: Iefa257bceb54137f25534ccb6786c1d2e315931c
(cherry picked from commit e27ae38bebf4b57aa7563357a188e33d0e1e0be8)
Signed-off-by: Andras Timar <andras.timar@collabora.com>