calcRTL: allow section right edge to anchor...
to the left of other sections.

This is necessary in RTL mode in Calc (without reversing the
x-coordinate globally). For instance to place the column header(tile
section) in the left side of the corner header(row header).

The alternate approach would be to reverse w.r.t x coordinate the
processingOrder's of the sections and anchor them appropriately. But
this would need lot more work to make the auto expand sections expand
correctly if they need to be 'processed'/'located' before others! As
currently there are no automated tests for canvas sections, it would
also need a lot of testing to cover all cases to check for regressions.
Hence the current approach was chosen which only introduce a new option
'-left' for relative anchoring.

Signed-off-by: Dennis Francis <dennis.francis@collabora.com>
Change-Id: Id6f0f65e3d1fbf637ba452042a8a7dc674339d7e