Make some shortcuts on Mac clients match expectations better
The expected shortcuts are documented in
https://support.apple.com/en-us/HT201236 . This patch handles a small
subset of those:

Command–Up Arrow: Move the insertion point to the beginning of the
document.

Command–Down Arrow: Move the insertion point to the end of the
document.

Command–Left Arrow: Move the insertion point to the beginning of the
current line.

Command–Right Arrow: Move the insertion point to the end of the
current line.

Option–Left Arrow: Move the insertion point to the beginning of the
previous word.

Option–Right Arrow: This should move the insertion point to the end of
the next word. But I don't think LO has such a shortcut so we move to
the beginning of the next word instead.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I6b0d36b0842388f0d49adf9b75fa9480df79c805