Calc: Add hover state to DropDownMarker and fix icon
Add missing hover state to DropDownMarker¹ and do not rely solely on the
SVG for all the styles. Instead, pass some to the HTML element so it can
be easily changed depending on its state (and also it allows in the
future for easy replacement of those colors with CSS Vars).

Fix icon:
- Use proper colors instead of default black
- 2 faced icon

1. Marker used as a button to trigger validity related menus