RTL: Serve html with correct RTL settings
In case of RTL users we want to switch UI to that mode
early so we don't see flipping in the browser.
Setup correct setting at the serving html step.
Using that approach we don't need to extend LOK API.

We detect that using "lang" parameter passed in the request.