revert: add edit option in link dialog.
At least in current master it is possible to right click on the link and
select 'Edit hyperlink' context menu item and the edit dialog(JS)
appears. So it seems there is no need for this workaround edit button
inside the warning 'JS' dialog box anymore.

Additional reason is that when help button is clicked on any
dialog(tunnelled or not) it shows a warning 'link' dialog that it is
about to open a external url, but this dialog has an unneeded edit
button because of the original commit.