jsdialog: use correct element to close popup
For toolitems with dropdowns we should toggle the button.
Use correct parent so we don't use button from other component.
Later we send toggle event to properly close the popup on the server.

example: Calc sidebar, Underline popup
Before this patch when we clicked in outside the popup it dissapeared
but app was completly blocked.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I124a41361df305931c66047a8799ce03393b67de