Avoid using undefine in custom button insertion
If postmessage to insert custom button will arrive
early before we initialized notebookbar it could
fail in insertion into compact mode toolbar due to
not existing toolbar.

This prevents us from that error. It will be not added
at the time of postmessage execution
but we remember all the custom buttons in special
array, so when UI will be initialized it will be added there.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I07d86db49a50be70b0c091a66cf5e4be049211ce