safari: fix contenteditable on old versions
In old versions of Safari, typing text would not input anything. This
appears to be a regression from replacing the input area with a
contenteditable div (1d2607e8363dab839e7060ab38612123d284c98f).

This seems to be caused by "Safari has the user-select CSS setting as
none by default" (see https://stackoverflow.com/a/31831776/12293760).
From my research, this only affects old versions of Safari, but adding
this CSS fixes it and does not prevent editing in other browsers.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I708708b87f3169f6f69cab28003379519ce114dd