Don't make view inactive when dialog is opened
If user opened any dialog and we got blur event on
map - it was doing the view inactive by sending
userinactive message to the server and ignoring any
meesages we tried to send. Because of that it was
impossible to test in cypress dropdown dismiss by
click outside popup (due to losing focus from map).

This is regression from rework:
commit 86e8491707c8f25b106909a340f45edd65e704e5
For .eslintrc change, see: https://github.com/typescript-eslint/typescript-eslint/issues/1824

where we changed:
- if (!this._active || isAnyVexDialogActive()) {
+ if (!this._active || isAnyVexDialogActive() || this.map.jsdialog.hasDialogOpened()) {

Make more precise condition so we check if the dialog is a dim one.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: Ic200bd3f0327e4c43d80fb8cf7c2fbd176485b82