jsdialog: openmenu action support in notebookbar
In notebookbar we insert additional containers for rows and cells.
When sending openmenu action we need to pass id of the real parent
of a button which is a toolbox. Go up in the structure to get
correct node.

This is 23.05 only as in master we already have simplified structure.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I070c700bf420ed6f1361ca32f88ad626e32dec84