calc: on tab switching document view can be restored to a wrong position
In calc when a tab is moved to a new position or a sheet is inserted
or deleted, saved positions in SheetSwitchViewRestore which are used
to restore the correct document.
position when user switches back to a tab are no more valid.
In fact the internal map: sheet index -> sheet position becomes wrong.
This patch takes care to update such internal map when one of the
listed event occurs.

Signed-off-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Change-Id: Ia4587b14179a35ff369d1f030657c4e5c64f5715