service: change KillMode to mixed
According to the systemd.kill man page, when
KillMode=control-group "all remaining processes
in the control group of this unit will be killed
on unit stop." This ends up killing the kit
processes as well, which prevents saving any
unsaved modifications before shutting down.

The setting we want is 'mixed'. According to
the man page, with mixed "the SIGTERM signal is
sent to the main process while the subsequent
SIGKILL signal is sent to all remaining processes
of the unit's control group."

In other words, with KillMode=mixed, oxoolwsd
will receive the termination signal and it will
manage its children. But systemd will still
make sure that when oxoolwsd terminates, or when
it times out, all children and properly killed.

This changes KillMode to mixed.

Signed-off-by: Firefly <firefly@ossii.com.tw>
Change-Id: I39f3168b4ff88bf15b4d10e14869ab25cbbc28fc