Fix video playback with load balancer
It's important to have correct WOPISrc for load balancers.
We need to decode "&" before other params like ServerId and
Tag so load balancer will not use it as a part of WOPISrc.

In other case when I had 2 instances of OXOOL behind HAProxy
sometimes request to get the video was forwarded to the
wrong instance.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I98a3343955c32d1bed9cb85b6eae7ea154b9be03